backend/requirements.txt
flask==3.1.2
flask-cors==4.0.1
python-dotenv
requests
